Move rendering code to separate files
