+++ src/RatchetAndClank/DataViewExt.ts
@@ -375,107 +375,3 @@ export class DataViewExt<T extends ArrayBufferLike = ArrayBufferLike> extends Da
         URL.revokeObjectURL(url);
     }
 }
-
-type DataViewExtWithTracerOptions = DataViewExtOptions & {
-    /**
-     * Tracks which bytes have been read. Must be the same length as this view. Subviews will inherit it. Initialized automatically if not provided.
-     */
-    tracer?: DataViewExt,
-};
-
-/**
- * A DataView that tracks which bytes have been read. Can be used to erase unwanted data from files to improve compression.
- */
-export class DataViewExtWithTracer<T extends ArrayBufferLike = ArrayBufferLike> extends DataViewExt<T> {
-    public tracer: DataViewExt;
-    constructor(buffer: T, options: DataViewExtWithTracerOptions, byteOffset?: number, byteLength?: number) {
-        super(buffer, options, byteOffset, byteLength);
-
-        this.tracer = options.tracer ?? new DataViewExt(new ArrayBuffer(this.byteLength), { littleEndian: this.defaultLittleEndian });
-        if (this.tracer.buffer.byteLength !== this.buffer.byteLength) {
-            throw new Error("Tracer's underlying buffer must be the same length as this view's underlying buffer");
-        }
-    }
-
-    private markRead(offset: number, length: number) {
-        for (let i = 0; i < length; i++) {
-            this.tracer.setUint8(offset + i, 1);
-        }
-    }
-
-    /**
-     * Set all bytes in this view to 0 if they have not been read.
-     */
-    eraseBytesNotMarkedAsRead() {
-        for (let i = 0; i < this.tracer.byteLength; i++) {
-            if (this.tracer.getUint8(i) === 0) {
-                this.setUint8(i, 0);
-            }
-        }
-    }
-
-    override getBigInt64(offset: number, littleEndian: boolean = this.defaultLittleEndian) {
-        this.markRead(offset, 8);
-        return super.getBigInt64(offset, littleEndian);
-    }
-
-    override getBigUint64(offset: number, littleEndian: boolean = this.defaultLittleEndian) {
-        this.markRead(offset, 8);
-        return super.getBigUint64(offset, littleEndian);
-    }
-
-    override getFloat16(offset: number, littleEndian: boolean = this.defaultLittleEndian) {
-        this.markRead(offset, 2);
-        return super.getFloat16(offset, littleEndian);
-    }
-
-    override getFloat32(offset: number, littleEndian: boolean = this.defaultLittleEndian) {
-        this.markRead(offset, 4);
-        return super.getFloat32(offset, littleEndian);
-    }
-
-    override getFloat64(offset: number, littleEndian: boolean = this.defaultLittleEndian) {
-        this.markRead(offset, 8);
-        return super.getFloat64(offset, littleEndian);
-    }
-
-    override getInt16(offset: number, littleEndian: boolean = this.defaultLittleEndian) {
-        this.markRead(offset, 2);
-        return super.getInt16(offset, littleEndian);
-    }
-
-    override getInt8(offset: number) {
-        this.markRead(offset, 1);
-        return super.getInt8(offset);
-    }
-
-    override getInt32(offset: number, littleEndian: boolean = this.defaultLittleEndian) {
-        this.markRead(offset, 4);
-        return super.getInt32(offset, littleEndian);
-    }
-
-    override getUint8(offset: number) {
-        this.markRead(offset, 1);
-        return super.getUint8(offset);
-    }
-
-    override getUint16(offset: number, littleEndian: boolean = this.defaultLittleEndian) {
-        this.markRead(offset, 2);
-        return super.getUint16(offset, littleEndian);
-    }
-
-    override getUint32(offset: number, littleEndian: boolean = this.defaultLittleEndian) {
-        this.markRead(offset, 4);
-        return super.getUint32(offset, littleEndian);
-    }
-
-    override subview(offset: number = 0, byteLength: number = 0) {
-        if (!byteLength) byteLength = this.byteLength - offset;
-        return new DataViewExtWithTracer(
-            this.buffer,
-            { littleEndian: this.defaultLittleEndian, tracer: this.tracer.subview(offset, byteLength) },
-            this.byteOffset + offset,
-            Math.min(byteLength, this.byteLength - offset)
-        );
-    }
-}
