+++ src/RatchetAndClank/bin-gameplay.ts
