+++ src/RatchetAndClank/tools/extractor.ts
@@ -4,13 +4,18 @@ import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import fs from "node:fs/promises";
 import { decompressWad } from "./decompress.ts";
-import { readLevelDataHeader, readLevelDescriptor, readTableOfContents } from "./structs-toc-and-level-headers.ts";
+import { readLevelDataHeader, readLevelDescriptor, readTableOfContents } from "../bin-toc.ts";
 import { readFromDisk, readFromDiskWithSizeHeader, SECTOR_SIZE } from "./utils.ts";
-import { DataViewExt, DataViewExtWithTracer } from "../DataViewExt.ts";
-import { readLevelCoreHeader } from "../structs-core.ts";
+import { DataViewExt } from "../DataViewExt.ts";
+import { readLevelCoreHeader } from "../bin-index.ts";
 import { assert } from '../../util.ts';
 
-const gameNumber = 1;
+const gameNumber = Number(process.argv[2]);
+
+if (gameNumber !== 1) {
+    console.error(`Usage: pnpm build:RatchetAndClank <gameNumber>`);
+    process.exit(1);
+}
 
 const baseDataDir = path.join(path.dirname(fileURLToPath(import.meta.url)), `../../../data`);
 
@@ -25,26 +30,16 @@ const disk = await fs.open(diskFile);
         - Global data (not used)
             - cutscenes, tutorials, loading screens, etc
         - Array of levels
-            - GS memory snapshot *output as level_n.gs_ram*
-            - Data file
-                - Core index *output as level_n.core_index*
-                    - Pointer tables, directory for core data
-                - Core data (compressed) *output as level_n.core_data after decompression*
-                    - Tfrag (a single world space mesh with 3 lods and lod tweening, used for terrain)
-                    - Sky
-                    - Classes (meshes with attached materials, lod config, bounding boxes, etc)
-                        - Ties (static meshes with 3 lods, used for rocks, trees, buildings, etc)
-                        - Mobys (meshes with skeletons and destructable parts, used for anything that moves)
-                        - Shrubs (meshes with a billboard lod, used for decorations)
-                    - Textures
+            - GS memory snapshot *output as level_n.gs*
+                - Palettes, some textures
+            - Core data index *output as level_n.index*
+                - Pointer tables, directory for core data
+            - Core data (compressed) *output as level_n.data after decompression*
+                - Models, textures, collision, etc
             - Gameplay file (compressed) *output as level_n.gameplay after decompression*
                 - Level config (fog, spawn point, etc)
                 - Instances of classes from core file
-                - Instances of collision shapes
-                - Direction lights
-                - Point lights
-            - Music (not used)
-            - Scenes (not used)
+                - Lights, paths, triggers, etc
 */
 
 const ENTRY_POINTS = {
@@ -69,9 +64,6 @@ for (const levelSectors of tableOfContents.levelSectors) {
     async function extractLevelFile(name: string, buf: string | Uint8Array | DataViewExt) {
         const filename = name.replace(/\{\}/g, String(levelNum));
         await fs.writeFile(path.join(outputDir, filename), buf);
-        if (buf instanceof DataViewExtWithTracer) {
-            await fs.writeFile(path.join(outputDir, filename + ".tracer"), buf.tracer);
-        }
         console.log(`Writing file ${filename}`);
         filesWritten.push(filename);
     }
@@ -82,35 +74,35 @@ for (const levelSectors of tableOfContents.levelSectors) {
     const levelData = new DataViewExt(levelDataBuffer, { littleEndian: true });
     const levelDataHeader = await readLevelDataHeader(levelData);
 
-    // level/gs_ram
+    // level/gs
     const gsRam = levelData.subview(levelDataHeader.gsRam.offset, levelDataHeader.gsRam.size);
 
     // level/gameplay
     const gameplaySector = levelDescriptor.gameplayNtsc;
     const gameplayFileCompressed = await readFromDiskWithSizeHeader(disk, gameplaySector.startSector, 0x3);
     const gameplayFileBuffer = decompressWad(new DataViewExt(gameplayFileCompressed, { littleEndian: true }));
-    const gameplayFile = new DataViewExtWithTracer(gameplayFileBuffer, { littleEndian: true });
+    const gameplayFile = new DataViewExt(gameplayFileBuffer, { littleEndian: true });
 
-    // level/core_index
+    // level/index
     const levelCoreIndex = levelData.subview(levelDataHeader.coreIndex.offset, levelDataHeader.coreIndex.size);
     const levelCoreHeader = await readLevelCoreHeader(levelCoreIndex);
 
-    // level/core_data
+    // level/core
     const levelCoreDataWad = levelData.subview(levelDataHeader.coreData.offset, levelDataHeader.coreData.size);
     assert(levelCoreDataWad.byteLength === levelCoreHeader.assetsCompressedSize);
     const levelCoreDataBuffer = decompressWad(levelCoreDataWad);
     assert(levelCoreDataBuffer.byteLength === levelCoreHeader.assetsDecompressedSize);
-    const levelCoreData = new DataViewExtWithTracer(levelCoreDataBuffer, { littleEndian: true });
+    const levelCoreData = new DataViewExt(levelCoreDataBuffer, { littleEndian: true });
 
     await extractLevelFile(`level_{}.gameplay`, gameplayFile);
-    await extractLevelFile(`level_{}.core_index`, levelCoreIndex);
-    await extractLevelFile(`level_{}.gs_ram`, gsRam);
-    await extractLevelFile(`level_{}.core_data`, levelCoreData);
+    await extractLevelFile(`level_{}.index`, levelCoreIndex);
+    await extractLevelFile(`level_{}.gs`, gsRam);
+    await extractLevelFile(`level_{}.core`, levelCoreData);
 
     // write metadata file
     await extractLevelFile(`level_{}.json`, JSON.stringify({
         filesReferenced: filesWritten,
         levelDataHeader,
         levelDescriptor,
-    }));
+    }, null, 2));
 }
